refactor: rename to hit_testing_base
